# public/swagger-specs/services/api/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "ConsenSys Codefi Orchestrate API. Enables dynamic management of transactions, identities, chains, faucets and contracts.\nTransaction Requests are an abstraction over schedules and jobs representing one or more transactions executed on the Blockchain network\nSchedules are ordered lists of jobs executed in a predefined sequence\nJobs represent single transaction flows executed on the Blockchain network\nChains represent list of endpoints pointing to a Blockchain network\nFaucets represent funded accounts (holding ETH) linked to specific chains, allowed to fund newly created accounts automatically for them to be able to send transactions.\nAccounts represent Ethereum accounts (private keys). By usage of the generated cryptographic key pair, accounts can be used to sign/verify and to encrypt/decrypt messages.\nContracts represent Solidity contracts management.\nEvent Streams represent Event streams management.",
        "title": "Orchestrate API",
        "contact": {
            "name": "Contact ConsenSys Codefi Orchestrate",
            "url": "https://consensys.net/codefi/orchestrate/contact",
            "email": "orchestrate@consensys.net"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "2.0"
    },
    "paths": {
...
        "/faucets": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    },
                    {
                        "JWTAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Faucets"
                ],
                "summary": "Retrieves a list of all registered faucets",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/types.FaucetResponse"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    },
                    {
                        "JWTAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Faucets"
                ],
                "summary": "Registers a new faucet",
                "parameters": [
                    {
                        "description": "Faucet registration request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/types.RegisterFaucetRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/types.FaucetResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "422": {
                        "description": "Unprocessable entity",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/faucets/{uuid}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Faucets"
                ],
                "summary": "Retrieves a faucet by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID of the faucet",
                        "name": "uuid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/types.FaucetResponse"
                        }
                    },
                    "404": {
                        "description": "Faucet not found",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    },
                    {
                        "JWTAuth": []
                    }
                ],
                "tags": [
                    "Faucets"
                ],
                "summary": "Deletes a faucet by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID of the faucet",
                        "name": "uuid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": ""
                    },
                    "400": {
                        "description": "Invalid request",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Faucet not found",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    }
                }
            },
            "patch": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    },
                    {
                        "JWTAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Faucets"
                ],
                "summary": "Updates a faucet by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID of the faucet",
                        "name": "uuid",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Faucet update request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/types.UpdateFaucetRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/types.FaucetResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Faucet not found",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "422": {
                        "description": "Unprocessable entity",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "api.ErrorResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 24000
                },
                "message": {
                    "type": "string",
                    "example": "error message"
                }
            }
        },
...
        "types.FaucetResponse": {
            "type": "object",
            "properties": {
                "amount": {
                    "description": "Amount, in Wei, sent to the beneficiary on each funding transaction.",
                    "type": "string",
                    "example": "0xD529AE9E860000"
                },
                "chainRule": {
                    "description": "Name of the chain on which the faucet is registered.",
                    "type": "string",
                    "example": "mainnet"
                },
                "cooldown": {
                    "description": "Waiting time in between two funding transactions to the same beneficiary.",
                    "type": "string",
                    "example": "10s"
                },
                "createdAt": {
                    "description": "Date and time at which the faucet was registered.",
                    "type": "string",
                    "example": "2020-07-09T12:35:42.115395Z"
                },
                "creditorAccount": {
                    "description": "Address of the faucet creditor's account.",
                    "type": "string",
                    "example": "0x905B88EFf8Bda1543d4d6f4aA05afef143D27E18"
                },
                "maxBalance": {
                    "description": "Expected maximum balance of beneficiary. It won't fund past this value.",
                    "type": "string",
                    "example": "0x16345785D8A0000"
                },
                "name": {
                    "description": "Name of the faucet.",
                    "type": "string",
                    "example": "faucet-mainnet"
                },
                "tenantID": {
                    "description": "ID of the tenant executing the API.",
                    "type": "string",
                    "example": "foo"
                },
                "updatedAt": {
                    "description": "Date and time at which the faucet details were updated.",
                    "type": "string",
                    "example": "2020-07-09T12:35:42.115395Z"
                },
                "uuid": {
                    "description": "UUID of the faucet.",
                    "type": "string",
                    "example": "b4374e6f-b28a-4bad-b4fe-bda36eaf849c"
                }
            }
        },
...
        "types.RegisterFaucetRequest": {
            "type": "object",
            "required": [
                "amount",
                "chainRule",
                "cooldown",
                "creditorAccount",
                "maxBalance",
                "name"
            ],
            "properties": {
                "amount": {
                    "description": "Amount, in Wei, sent to the beneficiary on each funding transaction.",
                    "type": "string",
                    "example": "0xF4240"
                },
                "chainRule": {
                    "description": "Name of the chain on which to register the faucet.",
                    "type": "string",
                    "example": "mainnet"
                },
                "cooldown": {
                    "description": "Waiting time in between two funding transactions to the same beneficiary.",
                    "type": "string",
                    "example": "10s"
                },
                "creditorAccount": {
                    "description": "Address of the faucet creditor's account.",
                    "type": "string",
                    "example": "0x905B88EFf8Bda1543d4d6f4aA05afef143D27E18"
                },
                "maxBalance": {
                    "description": "Expected maximum balance of beneficiary. It won't fund past this value.",
                    "type": "string",
                    "example": "0x254582f40"
                },
                "name": {
                    "description": "Name of the faucet.",
                    "type": "string",
                    "example": "faucet-mainnet"
                }
            }
        },
...
        "types.UpdateFaucetRequest": {
            "type": "object",
            "properties": {
                "amount": {
                    "description": "Amount, in Wei, sent to the beneficiary on each funding transaction.",
                    "type": "string",
                    "example": "0x254582f40"
                },
                "chainRule": {
                    "description": "Name of the chain on which to register the faucet.",
                    "type": "string",
                    "example": "mainnet"
                },
                "cooldown": {
                    "description": "Waiting time in between two funding transactions to the same beneficiary.",
                    "type": "string",
                    "example": "10s"
                },
                "creditorAccount": {
                    "description": "Address of the faucet creditor's account.",
                    "type": "string",
                    "example": "0x905B88EFf8Bda1543d4d6f4aA05afef143D27E18"
                },
                "maxBalance": {
                    "description": "Expected maximum balance of beneficiary. It won't fund past this value.",
                    "type": "string",
                    "example": "0x254582f40"
                },
                "name": {
                    "description": "Name of the faucet.",
                    "type": "string",
                    "example": "faucet-mainnet"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "X-API-Key",
            "in": "header"
        },
        "JWTAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}